requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.3.0
colorama==0.4.6
fastapi==0.116.1
h11==0.16.0
idna==3.10
packaging==25.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
pytesseract==0.3.13
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.23.2
